requirements.txt
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
dataclasses==0.6
deprecation==2.1.0
Django==3.2
django-cors-headers==3.7.0
django-extensions==3.1.3
django-filter==2.4.0
django-injector==0.2.2
django-model-utils==4.1.1
django-redis==4.12.1
djangorestframework==3.12.4
gotrue==0.2.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
idna==2.10
importlib-metadata==4.0.1
iniconfig==1.1.1
injector==0.18.4
packaging==20.9
pluggy==0.13.1
postgrest-py==0.4.0
py==1.10.0
pycryptodomex==3.10.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
realtime-py==0.1.2
redis==3.5.3
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
sqlparse==0.4.1
stream-python==5.1.0
supabase-py==0.0.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
websockets==8.1
zipp==3.4.1
